+++ NanoXLSX/LowLevel/XlsxWriter.cs
@@ -1212,7 +1212,12 @@ namespace NanoXLSX.LowLevel
             {
                 if (item.IsCustomFormat)
                 {
-                    sb.Append("<numFmt formatCode=\"").Append(item.CustomFormatCode).Append("\" numFmtId=\"").Append(item.CustomFormatID.ToString("G", culture)).Append("\"/>");
+                    if (string.IsNullOrEmpty(item.CustomFormatCode))
+                    {
+                        throw new FormatException("The number format style component with the ID " + Utils.ToString(item.CustomFormatID) + " cannot be null or empty");
+                    }
+                    String customFormat = NumberFormat.EscapeFormatCode(item.CustomFormatCode);
+                    sb.Append("<numFmt formatCode=\"").Append(EscapeXmlAttributeChars(customFormat)).Append("\" numFmtId=\"").Append(item.CustomFormatID.ToString("G", culture)).Append("\"/>");
                 }
             }
             return sb.ToString();